BitStamp: send an integer start in the candle request
ToUnix returns a double, so the request carried a fractional value and the venue
answered "Not a valid integer" - candle history never worked.

+++ BitStamp/Native/HttpClient.cs
@@ -54,7 +54,8 @@ class HttpClient : BaseLogReceiver
 
 		request.AddParameter("step", step);
 		request.AddParameter("limit", limit);
-		request.AddParameter("start", start.ToUnix());
+		// the endpoint rejects a fractional value with "Not a valid integer"
+		request.AddParameter("start", (long)start.ToUnix());
 		//request.AddParameter("end", end.ToUnix());
 		request.AddParameter("exclude_current_candle", true);
